FindByType values now match selenium's By and MobileBy values.
This allows creating a FindByType enum from By/MobileBy.

This feature is required for manually reporting executed addon steps.
The AddonHelper class has an execute method which requires an element's By (Selenium class) and value.
Manual reporting requires ElementSearchCriteria which is built from FindByType (built-in class of this SDK) and the value of the element.

Signed-off-by: Tzah Mazuz <tzah.mazuz@testproject.io>